Issue #8 - building on HAL
Changes to the JSON representation are minor - simply removing the array
of error object from the result.

The changes to the XML representation have changed more drastically as
the names of the elements themselves are now different. The XML variant
is more likely to be subject to change as the HAL XML spec has not been
finalised.